Link to new /doc/dev route
Summary:
This patch uses the new `/doc/dev` route instead of https://doc.bitcoinabc.org
everywhere in this repo.  A placeholder is introduced to ensure the route is not
accidentally reused since the dev docs are built and hosted separately from deployments
of the main site.

Test Plan:
```
make container
make run
```
Navigate, checking nav bar and the following links:
* http://127.0.0.1:8080/doc/ < should have a default index provided by nginx
* http://127.0.0.1:8080/doc/dev/ < should show the placeholder page

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7066