Add 0.21.12 RPC docs
Summary:
This adds the 0.21.12 RPC docs for the release going out today.

Note that the index.md files generated by the ABC build system were removed
since they are not yet production ready (they have no content, so we're letting
nginx take care of the indexing for us for now).

The index for these docs will be available at `https://bitcoinabc.org/doc/<version>/rpc`

Not all of the pages render well due to missing code blocks and other styling
factors, but we can fix these as we go since they render well enough that that
are not unusable.

In the future, we can automate adding the RPC docs to the site and maybe upload them
to a CMS or some other management system.

Depends on D7065 and D7066

Test Plan:
```
make container
make run
```
Navigate to http://127.0.0.1:8080/doc/0.21.12/rpc/

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7067