Add a link to RPC docs
Summary:
This adds a link for users to navigate to the new RPC docs.
The link currently points to an nginx-generated index.
The index also contains a /doc/dev link, which is confusing, but that
will go away when we start generating correct index pages.

Depends on D7067

Test Plan:
```
make container
make run
```
Navigate to: Developers -> RPC Documentation

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7068