Add top-level documentation index for serving docs by version
Summary:
This patch adds a top-level index for `/doc` and reduces
a click for navigating to any particular doc version.

Currently, the only versioned docs are the RPC docs, so the links
point to those directly.

The heading font sizes needed to be adjusted because they do not display well as-is.

Depends on D7130

Test Plan:
```
make serve
```
Navigate to http://localhost:4000/doc/

Also navigate around the site and make sure the css changes do not cause
any issues.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7131