Fix permalink in top-level doc index
Summary:
D7131 was only tested with `make serve` which runs a Jekyll
development server.  This server has additional routing rules which
makes the top-level index work as expected.  However, in production we
render the entire site to static files and the routing breaks. This is
due to an incorrect permalink, which is fixed by this patch.

Test Plan:
```
make container
make run
```
Navigate to http://localhost:8080/doc/ and verify the page loads as expected

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7179