Fix codeblock styling
Summary:
Current codeblock style has the following flaws:
* Distractly dark and bold line on the left side.
* Incorrectly aligned background coloring.

This patch fixes that as well as makes tweaks to padding and coloring
to be more pleasing to the eye.

This has the largest impact on large code blocks introduced by D7187

Depends on D7194 so that small devices do not see a strange line break
for some code blocks.

Test Plan:
`make serve`
Navigate to various pages with code blocks and verify they look good.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7195