[WEBSITE] Skip the RPC index during navigation
Summary:
The actual navigation flow for accessing the RPC documentation  is as follow:
{nav 1 - RPC documentation > 2 - Version selection > 3 - Chose the RPC command from a list (index page) > 4 - Navigate through the RPC commands (doc page)}

The index page at step 3 is not necessary since the documentation page from step 4 already contains a menu with all the commands accessible. It is just easier to jump to the documentation page and let the user select the command from there. That makes one less page with a dedicated layout to maintain for the same number of user clicks. New scheme:
{nav 1 - RPC documentation > 2 - Version selection > 3 - Navigate through the RPC commands (doc page)}

Test Plan: Build and test navigation through RPCs doc

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7250