[WEBSITE] Add a (user) documentation menu with the RPC documentation
Summary:
This diff moves the RPC documentation to a global user documentation
index page, reachable from the `Documentation` entry in the main menu.

The index design is updated accordingly, and now a table is used so that
other types of documentation can be easily added for each version as
needed. A new `category` variable will help distinguish the various doc
types, and default to the RPC docs.

Depends on D7250.

Test Plan: Build it and navigate from Documentation.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7251