[WEBSITE] Pull all markdown docs from ABC and render them
Summary:
This diff extends the `fetch-documentation.sh` script to copy all the
markdown files from the ABC repository to the website repository in
order to be processed and rendered by Jekyll.

The files can be accessed as if navigating through the ABC repository,
by using the project root relative link to the URL and substituting the
`.md` extension for `.html`. Examples:
`bitcoinabc.org/CONTRIBUTING.html`, `bitcoinabc.org/doc/build-unix.html`

In order to keep the links working between the markdown files, a couple
plugins are added to Jekyll:
 - The [jekyll-relative-links](https://github.com/benbalter/jekyll-relative-links)
   plugin will translate the relative `.md` links to their `.html`
   counterpart
 - A custom plugin will transpose to text the other relative links (such
   as source files) which cannot be rendered by Jekyll, and convert the
   project relative links (starting with a `/`) to file relative links
   so they are processed by `jekyll-relative-links`

The next step will be to replace the links to external markdown files
such as CONTRIBUTING.md and DISCLOSURE_POLICY.md with links to the local
pages.

Test Plan:
  make doc
  make serve
Navigate through the pages (e.g.
http://127.0.0.1:4000/CONTRIBUTING.html) and check there is no dead
link.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7437