[WEBSITE] Pick better defaults for RPC and man page documentations
Summary:
When navigating to the RPC or man page via the Documentation menu, after
the version selection the displayed page is the first in alphbetical
ordering from the first group.
This diff changes this behavior to put the `help` RPC and `bitcoind` man
page first if they exist, and fallback to previous behavior otherwise.

Test Plan: Build the website and navigate through the documentation.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7332