Remove unused avatar code
Summary:
None of our pages use the avatar image and it's questionable if we want this
to appear in the navbar. In order to not have to test unused code during the flexbox migration,
it's best to clean it up.

Test Plan:
`make serve`
Navigate around and ensure nothing is broken.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7404