Migrate all page content to flexbox
Summary:
This migrates the content structure of all pages to flexbox. As a result,
the subnav issues in the docs pages are fixed (no floating subnav over the footer).

It turns out that migrating the header and footer to flexbox is not strictly necessary
to complete before this patch.  However, we can still simplify the DOM and fix up a lot
of the styling by doing so.

Considering that this patch impacts all pages on the site, I request that all reviewers
actually run the test plan as it's not feasible for me to test on multiple versions of
multiple browsers.

Depends on D7399

Test Plan:
`make serve`
Navigate around and make sure nothing is broken using combinations of the following:
* Browser
  * Firefox
  * Chrome
  * etc...
* Screen size
  * small < 768px
  * medium < 900px
  * large > 900px
  * extra-large > 1200px
* Page type
  * [Homepage](http://localhost:4000/)
  * [RPC docs](http://localhost:4000/doc/0.22.0/rpc/blockchain/getblockchaininfo/)
  * [Manpages](http://localhost:4000/doc/0.22.0/man/bitcoin-wallet/)
  * [Announcements page](http://localhost:4000/announcements/)
    * Navigate to various posts on the announcements page

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7410