Fix logo link in non-prod environments
Summary:
Using `site.url` will always link to the production environment,
which is both wrong and confusing when deployed in beta.
Interestingly, deploying with `make serve` appears to override
to localhost.  The issue is only apparent in beta or when deployed
locally in a container.

Test Plan:
`make run`
Navigate to localhost:8080 and click the logo in the upper left. It should take you to localhost, instead of prod as it does pre-patch.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7412