Remove unused tags code
Summary:
Currently, we don't use tags for any of the posts on bitcoinabc.org.
Since we're using a CRM for blog.bitcoinabc.org, it seems unlikely we'll ever
actually use them, so the files and styles are just clutter.

Test Plan:
`make serve`
Navigate around, especially to http://127.0.0.1:4000/announcements/ and various
posts, taking note that nothing is broken.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D7431