Add release notes to the docs index
Summary:
Exposing the release notes on Github has been par for the course
for some time now, but we still get users asking where the release notes are.
Now that ABC markdown are exposed on the site, we should further expose the
release notes where the other docs are expected to be.

In the future, we can leverage the release notes to generate a release page, but
that will require some more work.

Test Plan:
```
scripts/fetch_documentation.sh
make serve
```
Navigate to http://127.0.0.1:4000/doc/

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7457