Fix blog feed escaping of unicode characters
Summary:
The pattern correctly matches the unicode characters, but
should not be saving the `u` character in the group. Otherwise the output
is something like `&#xu00e9;` which is invalid and ends up rendering as
those plain characters rather than the expected `é` which is `&#x00e9;`

Test Plan:
Remove the `.slice(0, 3)` from `js/blog-feed.js` to see more articles
if the recent ones do not display the above issue.
`make serve`
Navigate to localhost:4000 and see the feed rendering unicode characters correctly.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7465