extract markdown fetching from fetch_documentation.sh
Summary: It is convenient to run only the last part of fetch_documentation.sh in order to test any changes to the bitcoin-abc markdown files and how they affect the pages on the website. This diff moves the part of fetch_documentation.sh that is related to bitcoin-abc markdown to a separate script fetch_markdown_files.sh. This script takes the path to the local bitcoin-abc git repository as an argument, allowing to extract markdown files from a local developement repository?

Test Plan:
Check that `make doc` still works. Then delete the generated abc_md_docs directory.

Run `scripts/fetch_markdown_files.sh /path/to/bitcoin-abc` and check that it generated the abc_md_docs directory from the local repository (change something in any md file and confirm the change is present).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7497