Store releases JSON so the site can be version-aware
Summary:
By saving release JSON data to a file, we can access it during the build and
allow the site to be version-aware.

This patch provides only a proof-of-concept by introducing the current version
onto the homepage. In a future patch, I will leverage this further to build a
release page with both latest and recent versions.

Test Plan:
```
./fetch_documentation.sh
make serve
```
Navigate to localhost:4000 and observe the latest version number.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7574