Add version anchors to release page
Summary:
This allows us to link to specific releases (for as long as they are
available on the release page) using an anchor. Example: http://localhost:4000/releases/#0.22.0

This ensures that tweets and other social media posts have good SEO by not always pointing
to the same link (plain `/releases/`) over and over, as well as a better user experience
for people looking for old releases.

The implementation is a little unfortunately hacky since there is no way to define offsets for
an anchor. We need to make room for the nav bar, and anchoring the target content to the absolute
top edge of the screen is not always good UX anyway.

Test Plan:
`make serve`

Navigating to latest version brings you to the top of the page: http://localhost:4000/releases/#0.22.2

Navigating to most recent past version puts thoe whole past versions table in view: http://localhost:4000/releases/#0.22.1

Navigating to some older versions always puts them in view:
http://localhost:4000/releases/#0.22.0
http://localhost:4000/releases/#0.21.13
http://localhost:4000/releases/#0.21.2

Navigating to a very old version lands you on the top of the page: http://localhost:4000/releases/#0.20.0
(IMO this is good, as very old versions will tend to be out of consensus anyway, so encouraging the user to
upgrade should be the goal)

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7675