Error out if jq is not installed when fetching docs
Summary:
While the error message that a particular dependency is usually plain
to see when running this script, this is not necessarily true since we echo and
pipe long strings to jq, resulting in `jq: command not found` being
above-the-fold in these cases. We can fix this by providing an error message
to the user.

Test Plan:
  ./scripts/fetch_documentation.sh

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8127