Make a temporary copy of the latest release notes if the current version is not archived
Summary:
Currently, release notes are archived sometime after releases are tagged. With
our current process of letting a release candidate settle for a couple days, this is not
an issue. Howevever, rapid releases can cause bitcoinabc.org builds to break since the
latest archived release notes are not available.

We can fix this by copying the latest release notes file to where the archived release
notes are expected to be. This is not ideal, as initiating a release should be a more
atomic process, but until we have single-push-button releases, this will do for now.

Test Plan:
Default should continue to work without issue:
  ./scripts/fetch_documentation.sh
  make serve

Make the following edits to test a scenario where a new release does not have
archived release notes:
```
diff --git a/scripts/fetch_documentation.sh b/scripts/fetch_documentation.sh
index 6957d86..0623820 100755
--- a/scripts/fetch_documentation.sh
+++ b/scripts/fetch_documentation.sh
@@ -203,7 +203,7 @@ ${SCRIPT_DIR}/fetch_markdown_files.sh "${SRC_DIR}"
 # latest release notes available for that version number. This temporary copy
 # will be replaced once the release notes are archived.
 ABC_MD_DOCS="${TOPLEVEL}"/abc_md_docs
-LATEST_RELEASE_NOTES="${ABC_MD_DOCS}/doc/release-notes/release-notes-${RELEASE_VERSIONS[0]}.md"
+LATEST_RELEASE_NOTES="${ABC_MD_DOCS}/doc/release-notes/release-notes-0.22.7.md"
 if [ ! -f "${LATEST_RELEASE_NOTES}" ]; then
   cp "${ABC_MD_DOCS}/doc/release-notes.md" "${LATEST_RELEASE_NOTES}"
 fi
```
And copy the first top-level element in the JSON array in `_data/github-releases.json` and replace
`0.22.6` with `0.22.7`. Then run:
  ./scripts/fetch_documentation.sh
  make serve

The release page should display the current release notes without issue.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D8396