[WEBSITE] Fix fetching the latest release notes
Summary:
The `fetch_documentation.sh` script will make sure to link the latest
release notes file if it was not archived yet. But there is a bug in the
linked version number, as it links with the oldest version number
instead of the most recent one. This causes the website deployment to
fail finding the latest release notes, and prevent from displaying the
last one until the release notes gets archived.
This diff fixes the issue by using the correct index for getting the
latest version number from the list of available versions.

Test Plan:
  ./scripts/fetch_documentation.sh
  bundle exec jekyll serve
Check the latest version is 0.22.14.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9267