[website] remove exchanges that don't have withdrawals and deposits enabled
Summary:
As far as I know, this leaves only Coinex, STEX and Upbit for now.
The others have had plenty of time to re-enable deposits and withdrawals,
so we should remove them for now. We can re-add if/when they re-enable.

Test Plan: Built and locally and looked at the page and tested links.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9473