Add a notice of eCash rebranding on the release page
Summary:
Although the current 0.23.6 release notes include a mention of the `-ecash` option,
this will disappear from the release page once a new release goes out. This notice ensures
that node operators who are not entirely in the loop will be aware that the latest release
is eCash-compatible.

Test Plan:
  bundle exec jekyll serve

Navigate to http://localhost:4000/releases/ and review the notice message.

Reviewers: #bitcoin_abc, josephroyking

Reviewed By: #bitcoin_abc, josephroyking

Differential Revision: https://reviews.bitcoinabc.org/D9693