update /electrum page to eCash
Summary:
This is a minimalistic change to remove mentions of BCHA and BCHN, and
replace them with eCash and Bitcoin Cash.

There is still a link to the /bcha page, because of the relevant
information about the coin splitting and lack of replay protection.

Test Plan: Check the result locally with `bundle exec jekyll serve`

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9712