[website] Add FAQ question about burning
Summary:
People keep asking this, so let's answer it in the FAQ
Answer based on Amaury's answer from the AMA, plus expanded to explain why
XEC is actually very limited in supply in terms of future issuance.

Depends on D10217

Test Plan: Built it locally, tested link. Read it through to see that it makes sense.

Reviewers: bytesofman, #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10218