[WEBSITE} Limit the depth of the release documentation table
Summary: There is currently no limit on the number of release we display the documentation for. This diff limits to 15, which is in par with the release page.

Test Plan:
Get all the doc as needed. You can use the `scripts/fetch_documentation.sh` script or manually copy the folder from a version with another name in `_doc` to easily create fake data.
  bundle exec jekyll serve
Navigate to `Documentation` and check the limit is effective

Reviewers: #bitcoin_abc, Mengerian

Reviewed By: #bitcoin_abc, Mengerian

Differential Revision: https://reviews.bitcoinabc.org/D11060