[WEBSITE] Unbreaks the markdown links in the documentation
Summary:
D7615 introduced the releasepage that embeds the release notes markdown into the page. In order to do so, it renamed all the documentation files processed to add the jekyll front matter, which in turn broke all the link conversion because the permalink no longer match the file.
This diff reverses this behavior so the processed files remains with their original name as expected by the jekyll-relative-links plugin, and rename the unprocessed original instead. These original files can be embedded using their *.orig.md name where needed (only in the release page for now) and let alone all the other docs.

Test Plan:
  ./scripts/fetch_documentation.sh
  bundle exec jekyll serve
Navigate to "Developers=>Contributing" and check in the page that the link to the developer notes is now working as expected.
Check the release page is untouched.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11108