[WEBSITE] Properly convert the links in the release page
Summary:
The release page embeds a copy of the release notes markdown if the file has not been archived yet (which is always the case with our release process). If this release notes has a markdown link, it will not be converted by our plugins because only the files with a front matter are considered. IN order for these links to show correctly, we must convert them manually, which is easy enough because we know the source and destination directories.

Note that the release not will work fine after it is archived, because then it is linked from the html version and converted appropriately (assuming the link is correct in the first place).

Depends on D11108.

Test Plan:
  ./scripts/fetch_documentation.sh
  bundle exec jekyll serve
Check in the release page (for 0.25.0) that the i2p.md link is now working.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11109