[WEBSITE] Avoid null assignment when there is no link to convert in the release notes
Summary: Due to `set -u` in the `fetch_documentation.sh`, the script fails when there is no relative link in the release notes file because `REL_LINKS` is set to null. Fix it with double quotes so it's set to an empty string/array as expected.

Test Plan:
  ./scripts/fetch_documentation.sh
Will fail on 0.25.1 before this patch.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11159