Add sane instructions for rvm setup
Summary:
rvm.io's install instructions are not clear and do not work out of the box unless you pick the correct starting point
out of 10+ possible options.

Added instructions to the README that are copy-pastable but still link to rvm's original instructions just in case.

Test Plan: Try it yourself on a fresh machine or user.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D11398