[website] Improve "wen burn" FAQ answer
Summary:
The previous answer implied that "burning" supply is an option that the team just chooses not to do.
Clarify that the team doesn't control supply, in contrast to memetokens and VC coins.

Test Plan:
Read the answer, use judgement to assess if it addresses the common questions on eCash groups.
Ran the page locally to see that it renders nicely.

Reviewers: #bitcoin_abc, Kronkmeister

Reviewed By: #bitcoin_abc, Kronkmeister

Subscribers: Kronkmeister

Differential Revision: https://reviews.bitcoinabc.org/D11429