[website] Improve post-upgrade banner info
Summary:
Add link to Upgrade page, as it is still relevant after the upgrade and has more detailed info.
Add link to Releases page, as "below" isn't always relevant depending on which page the banner appears on (eg. on the /ecash page).

Test Plan: Ran it locally and checked links.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D11480