[website] Don't switch the upgrade banner status on the upgrade timestamp
Summary:
The upgrade usually happens 1h after the timestamp due to being MTP based. Turning the banner to the upgraded status too early is confusing, so let's give some margin instead. 3h is chosen as it gives enough time for a dev to change the value manually after the upgrade completed (or delay it if needed), and worst case it's just late by 2h.

Depends on D14621.

Test Plan:
  bundle exec jekyll serve --incremental
Check the upgrade notice is still displayed.
Manually change the switchTime to see it in effect.

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D14622